# golang/banking/src/handlers/api_client.go
/*
 * Swagger banking api simulation - OpenAPI 3.0
 *
 * Banking simulation application
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddClient - Add a new client
func AddClient(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// DepositMoney - Deposit money on client card
func DepositMoney(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetClient - get client by inn
func GetClient(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetJWT - Get a jwt token by providing phone and password
func GetJWT(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// TransferMoney - Transfer money to another client
func TransferMoney(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// UpdateClient - Update an existing client
func UpdateClient(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// WithdrawMoney - Withdraw money from client
func WithdrawMoney(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}