# golang/banking/src/handlers/api_credit.go
/*
 * Swagger banking api simulation - OpenAPI 3.0
 *
 * Banking simulation application
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// AddCredit - Create new credit
func AddCredit(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetCredit - get credit by id
func GetCredit(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// GetCredits - Get client's credits
func GetCredits(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}