# golang/banking/src/models/model_card.go
/*
 * Swagger banking api simulation - OpenAPI 3.0
 *
 * Banking simulation application
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type Card struct {

	CardId int64 `json:"cardId,omitempty"`

	Balance float64 `json:"balance,omitempty"`

	ValidTo string `json:"validTo,omitempty"`

	CvcCode string `json:"cvcCode,omitempty"`

	CardType string `json:"cardType,omitempty"`

	Currency string `json:"currency,omitempty"`

	ClientId int64 `json:"clientId,omitempty"`
}