# golang/banking/src/models/model_card_update.go
/*
 * Swagger banking api simulation - OpenAPI 3.0
 *
 * Banking simulation application
 *
 * API version: 1.0.11
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type CardUpdate struct {

	ValidTo string `json:"validTo,omitempty"`

	CvcCode string `json:"cvcCode,omitempty"`
}